+++ src/HotChocolate/Core/src/Execution/Processing/Plan/QueryPlanBuilder.cs
@@ -1,8 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Text.Json;
 using HotChocolate.Language;
 
 namespace HotChocolate.Execution.Processing.Plan
@@ -252,230 +249,5 @@ namespace HotChocolate.Execution.Processing.Plan
             selection.Strategy == SelectionExecutionStrategy.Serial
                 ? ExecutionStrategy.Serial
                 : ExecutionStrategy.Parallel;
-    }
-
-    internal sealed class ResolverQueryPlanNode : QueryPlanNode
-    {
-        public ResolverQueryPlanNode(ISelection first, ISelection? firstParent = null)
-            : base(QueryPlanBuilder.GetStrategyFromSelection(first))
-        {
-            First = first;
-            FirstParent = firstParent;
-            Selections.Add(first);
-        }
-
-        public ISelection? FirstParent { get; }
-
-        public ISelection First { get; }
-
-        public List<ISelection> Selections { get; } = new();
-
-        public override QueryPlanStep CreateStep()
-        {
-            var selectionStep = new ResolverQueryPlanStep(Strategy, Selections);
-
-            if (Nodes.Count == 0)
-            {
-                return selectionStep;
-            }
-
-            if (Nodes.Count == 1)
-            {
-                return new SequenceQueryPlanStep(
-                    new[]
-                    {
-                        selectionStep,
-                        Nodes[0].CreateStep()
-                    });
-            }
-
-            return new SequenceQueryPlanStep(
-                new QueryPlanStep[]
-                {
-                    selectionStep,
-                    new SequenceQueryPlanStep(Nodes.Select(t => t.CreateStep()).ToArray())
-                });
-        }
-
-        public override void Serialize(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WriteString("type", "Resolver");
-            writer.WriteString("strategy", Strategy.ToString());
-
-            writer.WritePropertyName("selections");
-            writer.WriteStartArray();
-            foreach (var selection in Selections)
-            {
-                writer.WriteStartObject();
-                writer.WriteNumber("id", selection.Id);
-                writer.WriteString("field", $"{selection.DeclaringType.Name}.{selection.Field.Name}");
-                writer.WriteString("responseName", selection.ResponseName);
-
-                if (selection.Strategy == SelectionExecutionStrategy.Pure)
-                {
-                    writer.WriteBoolean("pure", true);
-                }
-
-                if (selection.Strategy == SelectionExecutionStrategy.Inline)
-                {
-                    writer.WriteBoolean("inline", true);
-                }
-
-                writer.WriteEndObject();
-            }
-            writer.WriteEndArray();
-
-            if (Nodes.Count > 0)
-            {
-                writer.WritePropertyName("nodes");
-                writer.WriteStartArray();
-                foreach (var node in Nodes)
-                {
-                    node.Serialize(writer);
-                }
-                writer.WriteEndArray();
-            }
-
-            writer.WriteEndObject();
-        }
-    }
-
-    internal sealed class SequenceQueryPlanNode : QueryPlanNode
-    {
-        public SequenceQueryPlanNode() : base(ExecutionStrategy.Serial)
-        {
-        }
-
-        public override QueryPlanStep CreateStep()
-        {
-            return new SequenceQueryPlanStep(Nodes.Select(t => t.CreateStep()).ToArray());
-        }
-
-        public override void Serialize(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WriteString("type", "Sequence");
-
-            if (Nodes.Count > 0)
-            {
-                writer.WritePropertyName("nodes");
-                writer.WriteStartArray();
-                foreach (var node in Nodes)
-                {
-                    node.Serialize(writer);
-                }
-                writer.WriteEndArray();
-            }
-
-            writer.WriteEndObject();
-        }
-
-        public static SequenceQueryPlanNode Create(params QueryPlanNode[] nodes)
-        {
-            var sequence = new SequenceQueryPlanNode();
-            foreach (QueryPlanNode node in nodes)
-            {
-                sequence.AddNode(node);
-            }
-            return sequence;
-        }
-    }
-
-    internal sealed class ParallelQueryPlanNode : QueryPlanNode
-    {
-        public ParallelQueryPlanNode() : base(ExecutionStrategy.Parallel)
-        {
-        }
-
-        public override QueryPlanStep CreateStep()
-        {
-            return new SequenceQueryPlanStep(Nodes.Select(t => t.CreateStep()).ToArray());
-        }
-
-        public override void Serialize(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WriteString("type", "Parallel");
-
-            if (Nodes.Count > 0)
-            {
-                writer.WritePropertyName("nodes");
-                writer.WriteStartArray();
-                foreach (var node in Nodes)
-                {
-                    node.Serialize(writer);
-                }
-                writer.WriteEndArray();
-            }
-
-            writer.WriteEndObject();
-        }
-
-        public static ParallelQueryPlanNode Create(params QueryPlanNode[] nodes)
-        {
-            var parallel = new ParallelQueryPlanNode();
-            foreach (QueryPlanNode node in nodes)
-            {
-                parallel.AddNode(node);
-            }
-            return parallel;
-        }
-    }
-
-    internal abstract class QueryPlanNode
-    {
-        private readonly List<QueryPlanNode> _nodes = new();
-
-        protected QueryPlanNode(ExecutionStrategy strategy)
-        {
-            Strategy = strategy;
-        }
-
-        public ExecutionStrategy Strategy { get; }
-
-        public QueryPlanNode? Parent { get; private set; }
-
-        public IReadOnlyList<QueryPlanNode> Nodes => _nodes;
-
-        public void AddNode(QueryPlanNode node)
-        {
-            if (node is null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-
-            node.Parent = this;
-            _nodes.Add(node);
-        }
-
-        public void RemoveNode(QueryPlanNode node)
-        {
-            if (node is null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-
-            node.Parent = null;
-            _nodes.Remove(node);
-        }
-
-        public bool TryTakeNode([MaybeNullWhen(false)] out QueryPlanNode node)
-        {
-            if (_nodes.Count > 0)
-            {
-                node = _nodes[0];
-                node.Parent = null;
-                _nodes.RemoveAt(0);
-                return true;
-            }
-
-            node = null;
-            return false;
-        }
-
-        public abstract QueryPlanStep CreateStep();
-
-        public abstract void Serialize(Utf8JsonWriter writer);
     }
 }
